implement initial sync + refactor download to utils
